test: updated snapshots
