+++ src/components/Navigation/DesktopNavigationItem.test.js
@@ -1,9 +1,34 @@
 import React from "react";
 import renderer from "react-test-renderer";
-import HorizontalNavigationItem from "./HorizontalNavigationItem";
+import DesktopNavigationItem from "./DesktopNavigationItem";
+
+const testMenuItem = {
+  _id: "123",
+  name: "Shop",
+  slug: "shop"
+};
+
+const testMenuItemWithRelatedTags = {
+  _id: "123",
+  name: "Shop",
+  slug: "shop",
+  relatedTags: [
+    { _id: "111", name: "Clothes", slug: "clothes" },
+    { _id: "222", name: "Shoes", slug: "shoes" },
+    { _id: "333", name: "Accessories", slug: "accessories" }
+  ]
+};
 
 test("basic snapshot", () => {
-  const component = renderer.create(<HorizontalNavigationItem />);
+  const component = renderer.create(<DesktopNavigationItem menuItem={testMenuItem} />);
+
+  const tree = component.toJSON();
+  expect(tree).toMatchSnapshot();
+});
+
+test("snapshot with related items", () => {
+  const component = renderer.create(<DesktopNavigationItem menuItem={testMenuItemWithRelatedTags} />);
+
   const tree = component.toJSON();
   expect(tree).toMatchSnapshot();
 });
